refactor: Remove circular dependencies
