+++ packages/mermaid/src/docs/package.json
@@ -17,7 +17,8 @@
   "dependencies": {
     "@vueuse/core": "^10.1.0",
     "jiti": "^1.18.2",
-    "vue": "^3.3"
+    "vue": "^3.3",
+    "mermaid": "workspace:^"
   },
   "devDependencies": {
     "@iconify-json/carbon": "^1.1.16",
